+++ examples/function/multiproc_map/pyproject.toml
@@ -6,7 +6,7 @@ authors = ["Numaflow developers"]
 
 [tool.poetry.dependencies]
 python = "~3.10"
-pynumaflow = "~0.3.2"
+pynumaflow = "~0.4.1"
 
 [tool.poetry.dev-dependencies]
