+++ pynumaflow/tests/function/test_multiproc.py
@@ -5,7 +5,7 @@ from unittest import mock
 import grpc
 
 from pynumaflow.function import UserDefinedFunctionServicer
-from pynumaflow.function.multiproc_server import MultiProcServer
+from pynumaflow.function._multiproc_server import MultiProcServer
 from pynumaflow.function.proto import udfunction_pb2_grpc
 from pynumaflow.tests.function.test_server import (
     mapt_handler,
